# src/main/resources/openapi/products-openapi.yml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Product Catalog API",
    "description": "A robust API for managing a product catalog, including inventory, pricing, and detailed product information.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.example.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api.staging.example.com/v1",
      "description": "Staging server"
    }
  ],
  "tags": [
    {
      "name": "Products",
      "description": "Endpoints for managing products."
    },
    {
      "name": "Categories",
      "description": "Endpoints for managing product categories."
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Retrieve a list of products.",
        "description": "Fetches a paginated list of products with optional filtering and sorting.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "The number of products to return per page.",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set.",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "Filter products by category ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "A search query to filter products by name or description.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "Total number of products matching the query."
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new product.",
        "description": "Adds a new product to the catalog.",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "description": "Product object to be created.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "Conflict, a product with this name or SKU already exists."
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "summary": "Retrieve a product by ID.",
        "description": "Fetches a single product by its unique ID.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The unique ID of the product.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "summary": "Update an existing product.",
        "description": "Updates all fields of a product by its ID.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The unique ID of the product.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Product object with updated data.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a product.",
        "description": "Deletes a product by its ID.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The unique ID of the product.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "productId",
          "name",
          "price",
          "sku"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The unique identifier for the product.",
            "readOnly": true,
            "example": "prod-4a2b8c9d-1e1f-4f7e-9b5d-0d6c8b9a1c2d"
          },
          "name": {
            "type": "string",
            "description": "The name of the product.",
            "example": "Wireless Ergonomic Mouse"
          },
          "description": {
            "type": "string",
            "description": "A detailed description of the product.",
            "example": "A comfortable wireless mouse designed for long hours of use."
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit. A unique identifier for the product variant.",
            "example": "WM-001-BLK"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The current price of the product.",
            "example": 49.99
          },
          "currency": {
            "type": "string",
            "description": "The currency of the price.",
            "default": "USD",
            "example": "USD"
          },
          "stock": {
            "type": "integer",
            "description": "The number of units in stock.",
            "example": 150
          },
          "categoryId": {
            "type": "string",
            "description": "The category the product belongs to.",
            "nullable": true,
            "example": "cat-a1b2c3d4e5"
          },
          "images": {
            "type": "array",
            "description": "A list of image URLs for the product.",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "example": [
              "https://example.com/images/prod-001-main.jpg",
              "https://example.com/images/prod-001-side.jpg"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the product was created.",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the product was last updated.",
            "readOnly": true
          }
        }
      },
      "NewProduct": {
        "type": "object",
        "required": [
          "name",
          "price",
          "sku"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the product.",
            "example": "Wireless Ergonomic Mouse"
          },
          "description": {
            "type": "string",
            "description": "A detailed description of the product.",
            "example": "A comfortable wireless mouse designed for long hours of use."
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit. A unique identifier for the product variant.",
            "example": "WM-001-BLK"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The current price of the product.",
            "example": 49.99
          },
          "currency": {
            "type": "string",
            "description": "The currency of the price.",
            "default": "USD",
            "example": "USD"
          },
          "stock": {
            "type": "integer",
            "description": "The number of units in stock.",
            "example": 150
          },
          "categoryId": {
            "type": "string",
            "description": "The category the product belongs to.",
            "nullable": true,
            "example": "cat-a1b2c3d4e5"
          },
          "images": {
            "type": "array",
            "description": "A list of image URLs for the product.",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "example": [
              "https://example.com/images/prod-001-main.jpg",
              "https://example.com/images/prod-001-side.jpg"
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A unique error code."
          },
          "message": {
            "type": "string",
            "description": "A human-readable error message."
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was malformed or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}